requirements.txt
streamlit
pandas
pymssql
azure-storage-blob
openpyxl
sqlalchemy